refactor add to playlist
