+++ src/App.tsx
@@ -1,13 +1,8 @@
 import React from "react";
-import { BrowserRouter } from "react-router-dom";
 import AppRouter from "./app.router";
 
 function App() {
-  return (
-    <BrowserRouter>
-      <AppRouter />
-    </BrowserRouter>
-  );
+  return <AppRouter />;
 }
 
 export default App;
