+++ src/index.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
+import { BrowserRouter } from "react-router-dom";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import { makeServer } from "./server";
@@ -10,10 +11,12 @@ makeServer();
 
 ReactDOM.render(
   <React.StrictMode>
-    <Provider>
-      <GlobalStyle />
-      <App />
-    </Provider>
+    <BrowserRouter>
+      <Provider>
+        <GlobalStyle />
+        <App />
+      </Provider>
+    </BrowserRouter>
   </React.StrictMode>,
   document.getElementById("root")
 );
