+++ src/pages/components/playlistModal/index.js
@@ -1,0 +1,3 @@
+import PlaylistModal from "./playlistModal";
+
+export default PlaylistModal;
