+++ src/pages/watchlater/watchlater.styled.js
@@ -17,4 +17,7 @@ export const Wrapper = styled.div`
       transform: translate(-50%, -50%);
     }
   }
+  .video {
+    min-width: 30rem;
+  }
 `;
